+++ rosomaxa/src/algorithms/math/mod.rs
@@ -2,3 +2,6 @@
 
 mod distance;
 pub use self::distance::*;
+
+mod statistics;
+pub use self::statistics::*;