+++ rosomaxa/tests/unit/heuristics/termination/max_generation_test.rs
@@ -1,0 +1,3 @@
+#[test]
+#[ignore]
+fn can_use_max_generation() {}