+++ rosomaxa/tests/unit/heuristics/termination/min_variation_test.rs
@@ -1,0 +1,3 @@
+#[test]
+#[ignore]
+fn can_use_min_variation() {}