+++ vrp-core/src/algorithms/statistics/mod.rs
@@ -1,4 +1,0 @@
-//! This module contains some statistic related functionality.
-
-mod basics;
-pub use self::basics::*;